Move logic into AllianceRequestExecution
